Move firestore tests (#7318)
