+++ src/functions/env.spec.ts
@@ -4,8 +4,8 @@ import * as os from "os";
 import { sync as rimraf } from "rimraf";
 import { expect } from "chai";
 
-import * as env from "../../functions/env";
-import { FirebaseError } from "../../error";
+import * as env from "./env";
+import { FirebaseError } from "../error";
 
 describe("functions/env", () => {
   describe("parse", () => {
